% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_algo_interface.R
\name{gf_algo}
\alias{gf_algo}
\title{Create a GraphFrame Algorithm Object}
\usage{
gf_algo(class, algo, ..., .call = sys.call(sys.parent()))
}
\arguments{
\item{class}{The name of the graph algorithm used. Note that the
model name generated will be generated as \code{ml_model_<class>};
that is, \code{gf_algo} will be prefixed.}

\item{algo}{The underlying GraphFrame algo object.}

\item{...}{Additional algorithm information; typically supplied as named
values.}

\item{.call}{The \R call used in generating this algo object (ie,
the top-level \R routine that wraps over the associated GraphFrame
routine). Typically used for print output in e.g. \code{print}
and \code{summary} methods.}
}
\description{
Create a GF algo object, wrapping the result of a GraphFrame routine call.
The generated object will be an \R list with S3 classes
\code{c("gf_algo_<class>", "gf_algo")}.
}
