% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_algo_interface.R
\name{gf_algo_result}
\alias{gf_algo_result}
\title{Extracts the result of a GraphFrame algo run}
\usage{
gf_algo_result(object, ...)
}
\arguments{
\item{object}{a \code{gf_algo} object}

\item{...}{additional arguments, currently not used}
}
\description{
Extracts the result of a GraphFrame algo run
}
