% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_examples.R
\name{gf_chain}
\alias{gf_chain}
\title{Chain graph}
\usage{
gf_chain(sc, n)
}
\arguments{
\item{sc}{A Spark connection.}

\item{n}{Size of the graph to return.}
}
\description{
Returns a chain graph of the given size with Long ID type.
  The vertex IDs are 0, 1, ..., n-1, and the edges are (0, 1), (1, 2), ...., (n-2, n-1).
}
\examples{
\dontrun{
gf_chain(sc, 5)
}
}
