% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_interface.R
\name{gf_degrees}
\alias{gf_degrees}
\title{Degrees of vertices}
\usage{
gf_degrees(x)
}
\arguments{
\item{x}{An object coercable to a GraphFrame (typically, a
\code{gf_graphframe}).}
}
\description{
Degrees of vertices
}
