% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_interface.R
\name{gf_edge_columns}
\alias{gf_edge_columns}
\title{Edges column names}
\usage{
gf_edge_columns(x)
}
\arguments{
\item{x}{An object coercable to a GraphFrame (typically, a
\code{gf_graphframe}).}
}
\description{
Edges column names
}
