% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_interface.R
\name{gf_edges}
\alias{gf_edges}
\title{Extract edges DataFrame}
\usage{
gf_edges(x)
}
\arguments{
\item{x}{An object coercable to a GraphFrame (typically, a
\code{gf_graphframe}).}
}
\description{
Extract edges DataFrame
}
