% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_interface.R
\name{gf_find}
\alias{gf_find}
\title{Motif finding: Searching the graph for structural patterns}
\usage{
gf_find(x, pattern)
}
\arguments{
\item{x}{An object coercable to a GraphFrame (typically, a
\code{gf_graphframe}).}

\item{pattern}{pattern specifying a motif to search for}
}
\description{
Motif finding: Searching the graph for structural patterns
}
