% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_examples.R
\name{gf_friends}
\alias{gf_friends}
\title{Graph of friends in a social network.}
\usage{
gf_friends(sc)
}
\arguments{
\item{sc}{A Spark connection.}
}
\description{
Graph of friends in a social network.
}
\examples{
\dontrun{
library(sparklyr)
sc <- spark_connect(master = "local")
gf_friends(sc)
}
}
