% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_interface.R
\name{gf_graphframe}
\alias{gf_graphframe}
\title{Create a new GraphFrame}
\usage{
gf_graphframe(vertices = NULL, edges)
}
\arguments{
\item{vertices}{a tbl_spark representing vertices}

\item{edges}{a tbl_psark representing edges}
}
\description{
Create a new GraphFrame
}
