% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_interface.R
\name{gf_in_degrees}
\alias{gf_in_degrees}
\title{In-degrees of vertices}
\usage{
gf_in_degrees(x)
}
\arguments{
\item{x}{An object coercable to a GraphFrame (typically, a
\code{gf_graphframe}).}
}
\description{
In-degrees of vertices
}
