% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_lpa.R
\name{gf_lpa}
\alias{gf_lpa}
\title{Label propagation algorithm (LPA)}
\usage{
gf_lpa(x, max_iter, ...)
}
\arguments{
\item{x}{An object coercable to a GraphFrame (typically, a
\code{gf_graphframe}).}

\item{max_iter}{maximum number of iterations}

\item{...}{Optional arguments, currently not used.}
}
\description{
Label propagation algorithm (LPA)
}
