% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_interface.R
\name{gf_persist}
\alias{gf_persist}
\title{Persist the GraphFrame}
\usage{
gf_persist(x, storage_level = "MEMORY_AND_DISK")
}
\arguments{
\item{x}{An object coercable to a GraphFrame (typically, a
\code{gf_graphframe}).}

\item{storage_level}{The storage level to be used. Please view the
\href{http://spark.apache.org/docs/latest/programming-guide.html#rdd-persistence}{Spark Documentation}
for information on what storage levels are accepted.}
}
\description{
Persist the GraphFrame
}
