% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_interface.R
\name{gf_register}
\alias{gf_register}
\title{Register a GraphFrame object}
\usage{
gf_register(x)
}
\arguments{
\item{x}{An object coercable to a GraphFrame (typically, a
\code{gf_graphframe}).}
}
\description{
Register a GraphFrame object
}
