% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_scc.R
\name{gf_scc}
\alias{gf_scc}
\title{Strongly connected components}
\usage{
gf_scc(x, max_iter, ...)
}
\arguments{
\item{x}{An object coercable to a GraphFrame (typically, a
\code{gf_graphframe}).}

\item{max_iter}{maximum number of iterations}

\item{...}{Optional arguments, currently not used.}
}
\description{
Strongly connected components
}
