% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_shortest_paths.R
\name{gf_shortest_paths}
\alias{gf_shortest_paths}
\title{Shortest paths}
\usage{
gf_shortest_paths(x, landmarks, ...)
}
\arguments{
\item{x}{An object coercable to a GraphFrame (typically, a
\code{gf_graphframe}).}

\item{landmarks}{IDs of landmark vertices.}

\item{...}{Optional arguments, currently not used.}
}
\description{
Computes shortest paths from every vertex to the given set of landmark vertices.
  Note that this takes edge direction into account.
}
\examples{
\dontrun{
g <- gf_friends(sc)
gf_shortest_paths(g, landmarks = c("a", "d"))
}
}
