% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_examples.R
\name{gf_star}
\alias{gf_star}
\title{Generate a star graph}
\usage{
gf_star(sc, n)
}
\arguments{
\item{sc}{A Spark connection.}

\item{n}{The number of leaves.}
}
\description{
Returns a star graph with Long ID type, consisting of a central element
   indexed 0 (the root) and the n other leaf vertices 1, 2, ..., n.
}
\examples{
\dontrun{
gf_star(sc, 5)
}
}
