% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_triangle_count.R
\name{gf_triangle_count}
\alias{gf_triangle_count}
\title{Computes the number of triangles passing through each vertex.}
\usage{
gf_triangle_count(x, ...)
}
\arguments{
\item{x}{An object coercable to a GraphFrame (typically, a
\code{gf_graphframe}).}

\item{...}{Optional arguments, currently not used.}
}
\description{
This algorithm ignores edge direction; i.e., all edges are treated
  as undirected. In a multigraph, duplicate edges will be counted only once.
}
\examples{
\dontrun{
g <- gf_friends(sc)
gf_triangle_count(g)
}
}
