% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_examples.R
\name{gf_two_blobs}
\alias{gf_two_blobs}
\title{Generate two blobs}
\usage{
gf_two_blobs(sc, blob_size)
}
\arguments{
\item{sc}{A Spark connection.}

\item{blob_size}{The size of each blob.}
}
\description{
Two densely connected blobs (vertices 0->n-1 and n->2n-1)
  connected by a single edge (0->n).
}
\examples{
\dontrun{
gf_two_blobs(sc, 3)
}
}
