% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_interface.R
\name{gf_unpersist}
\alias{gf_unpersist}
\title{Unpersist the GraphFrame}
\usage{
gf_unpersist(x, blocking = FALSE)
}
\arguments{
\item{x}{An object coercable to a GraphFrame (typically, a
\code{gf_graphframe}).}

\item{blocking}{whether to block until all blocks are deleted}
}
\description{
Unpersist the GraphFrame
}
