% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_interface.R
\name{gf_vertex_columns}
\alias{gf_vertex_columns}
\title{Vertices column names}
\usage{
gf_vertex_columns(x)
}
\arguments{
\item{x}{An object coercable to a GraphFrame (typically, a
\code{gf_graphframe}).}
}
\description{
Vertices column names
}
