% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_interface.R
\name{gf_vertices}
\alias{gf_vertices}
\title{Extract vertices DataFrame}
\usage{
gf_vertices(x)
}
\arguments{
\item{x}{An object coercable to a GraphFrame (typically, a
\code{gf_graphframe}).}
}
\description{
Extract vertices DataFrame
}
