% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_interface.R
\name{spark_graphframe}
\alias{spark_graphframe}
\title{Retrieve a GraphFrame}
\usage{
spark_graphframe(x, ...)

spark_graphframe(x, ...)
}
\arguments{
\item{x}{An object coercable to a GraphFrame (typically, a
\code{gf_graphframe}).}

\item{...}{additional arguments, not used}
}
\description{
Retrieve a GraphFrame
}
