% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_scc.R
\name{gf_scc}
\alias{gf_scc}
\title{Strongly connected components}
\usage{
gf_scc(x, max_iter, ...)
}
\arguments{
\item{x}{An object coercable to a GraphFrame (typically, a
\code{gf_graphframe}).}

\item{max_iter}{Maximum number of iterations.}

\item{...}{Optional arguments, currently not used.}
}
\description{
Compute the strongly connected component (SCC) of each vertex and return a
  DataFrame with each vertex assigned to the SCC containing that vertex.
}
\examples{
\dontrun{
g <- gf_friends(sc)
gf_scc(g, max_iter = 10)
}
}
